+++ nexuse2e-webapp/src/main/frontend/src/app/header/header.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from "@angular/core";
-import { NavigationService } from "../navigation/navigation.service";
+import { NavigationService } from "../services/navigation.service";
 import { DataService } from "../services/data.service";
 
 @Component({
