+++ nexuse2e-webapp/src/main/frontend/src/app/services/navigation.service.ts
